Create README.md
Aim of this project is to write text on up to 4 neopixel matrices chained together. 

not all letters implemented yet in MAINX_short... which is the current test file. To write a text is has to be hardcoded in the run function. 

Aim is to write a text, in different colours, and let it scroll over the display.