Added a dummy of the neopixel lib
this one is ues for easier debugging the neotext dummy with CPython.

Added a first attempt for scrolling of the text with run_text().
But it is slow and does not allow to use more than 32 columns. Added
a destructor

Refactoring: new class COLORS as enum
and class Color as a common interface for color representation as
tupel and as integer